# [REVIEW] Fix slow decompression performance