--- a/cpp/src/io/utilities/block_utils.cuh
+++ b/cpp/src/io/utilities/block_utils.cuh
@@ -39,7 +39,7 @@ inline __device__ uint32_t ballot(int pred) { return __ballot_sync(~0, pred); }
 template <typename T>
 inline __device__ void nanosleep(T d)
 {
-  // TODO: nanosleep temporarily disabled 
+  // TODO: nanosleep temporarily disabled
   clock();
 }