Fixes #7114 

No performance regression on Turing (T4, CUDA 11.0, driver 450.102.04):
<details><summary>Click here to see environment details</summary><pre>
nvprof --print-gpu-trace ./gbenchmarks/PARQUET_READER_BENCH --benchmark_filter="ParquetRead/integral_file_input/28/1000/1/1/0/manual_time" 2>&1

*** WITH PATCH *** 
   Start  Duration            Grid Size      Block Size     Regs*    SSMem*    DSMem*      Size  Throughput  SrcMemType  DstMemType           Device   Context    Stream  Name
3.34085s  9.8091ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [4661]
4.77212s  9.6983ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [9822]
4.88598s  9.8804ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [10833]
5.00339s  9.9037ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [11844]
5.12055s  9.8821ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [12855]
5.23300s  9.8300ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [13866]
5.34499s  9.7688ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [14877]

*** WITHOUT PATCH ***

   Start  Duration            Grid Size      Block Size     Regs*    SSMem*    DSMem*      Size  Throughput  SrcMemType  DstMemType           Device   Context    Stream  Name
3.34647s  10.088ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [4661]
4.70031s  9.6986ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [9822]
4.80015s  9.9992ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [10833]
4.90408s  9.8958ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [11844]
5.00237s  9.8824ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [12855]
5.10425s  9.9476ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [13866]
5.20341s  9.8957ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [14877]
5.30279s  9.8980ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [15888]
</pre></details>

<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. There are CI checks in place to enforce that committed code follows our style
   and syntax standards. Please see our contribution guide in `CONTRIBUTING.MD`
   in the project root for more information about the checks we perform and how
   you can run them locally.

4. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

5. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please mark your pull request as Draft.
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#converting-a-pull-request-to-a-draft

6. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove it from "Draft" and make it "Ready for Review".
   https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/changing-the-stage-of-a-pull-request#marking-a-pull-request-as-ready-for-review

   If assistance is required to complete the functionality, for example when the
   C/C++ code of a feature is complete but Python bindings are still required,
   then add the label `help wanted` so that others can triage and assist.
   The additional changes then can be implemented on top of the same PR.
   If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

7. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on the target branch, force push, or rewrite history.
   Doing any of these causes the context of any comments made by reviewers to be lost.
   If conflicts occur against the target branch they should be resolved by
   merging the target branch into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->