@devavret 2021-04-14
Can you also post the results with CUDA 10.2? The nanosleep issue is in driver versions 450.80.02 to 450.110 but only visible in CUDA 10.2 because CUDA 11 nvcc patches it.