@revans2 2021-04-14
I tested this against the files that were causing issues on by dev box that caused me to file #7114. This fixes them.  The files that took 1.8 seconds before now finish in 90 ms, and a complete build of the spark plugin with tests went from 38 mins back to 30.