@revans2 2021-04-14
> @revans2 if you could test again with the latest changes it would appreciated as well!

I did a quick test with it and it looks good. I don't have the exact numbers right now, but I will try to get them