@elstehle 2021-04-14
> @elstehle Can you please post the performance after your current commit where you have removed nanosleep.

This time tested with CUDA 10.2 (driver : (T4, CUDA 10.2, driver 450.102.04)
```
*** PATCH IN PLACE***
ParquetRead/integral_file_input/28/1000/1/1/0/manual_time        116 ms          116 ms            6 bytes_per_second=4.32857G/s

nvprof --print-gpu-trace ./gbenchmarks/PARQUET_READER_BENCH --benchmark_filter="ParquetRead/integral_file_input/28/1000/1/1/0/manual_time" 2>&1|grep unsnap
2.83721s  14.245ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [4657]
4.24138s  13.080ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [9818]
4.35756s  13.359ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [10829]
4.47735s  13.361ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [11840]
4.59230s  13.424ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [12851]
4.70582s  13.407ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [13862]
4.82084s  13.386ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [14873]


*** NO PATCH***
ParquetRead/integral_file_input/28/1000/1/1/0/manual_time        120 ms          120 ms            5 bytes_per_second=4.15742G/s

nvprof --print-gpu-trace ./gbenchmarks/PARQUET_READER_BENCH --benchmark_filter="ParquetRead/integral_file_input/28/1000/1/1/0/manual_time" 2>&1|grep unsnap
2.79920s  17.970ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [4657]
4.22674s  15.256ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [9818]
4.34504s  15.410ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [10829]
4.46671s  16.608ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [11840]
4.58445s  15.099ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [12851]
4.70102s  16.133ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [13862]
4.81767s  16.921ms            (954 1 1)       (128 1 1)        44  1.5938KB        0B         -           -           -           -     Tesla T4 (0)         1         7  void cudf::io::unsnap_kernel<int=128>(cudf::io::gpu_inflate_input_s*, cudf::io::gpu_inflate_status_s*) [14873]
```