@revans2 2021-04-14
Yup the numbers are about the same.  about 93 ms for reading the file that used to take 1.8 seconds and 31 mins to run the full set of tests instead of 38. Don't know if it is run to run variance or not, but either way it is a lot better.